Fix in the compiler (could loop on circular structures)
Compiling (cons #0=(a . #0#) #1=(a . #1#)) lead to an infinite loop
since it searches in contants using the member function . Using memq
solves the problem, at the expense of a little bigger tables.

Using (member x y equiv?) is not a good idea (until it is optimized)
since each bootstrap phase is more than 10 times slower.