Added meta informations of the source distrbution files in Git
Since Git doesn't store modification dates and changed file protections,
clones are not completely identical to the original source tree. This is
not a problem per se except that it obliges the cloner to have a version of
autotools installed. This is useful too to build source distributions
archives where Makefiles files must be more recent than their associated
Makefile.am to avoid their regeneration.

This is needed to permit to rebuild a source tree coidentical