Adds SRFI-173
This is the example implementation from the SRFI, adapted to STklos.

One procedure was changed: instead of "assume" in hook-run,
an if followed by a call to error is done. This is consistent
with what is done in other SRFI implementations in STklos.

The tests are simple, but all procedures are tested.