Adaptation of the threads for Android
Android lacks support for the `pthread_cancel` function. Consequently
the current implementation of `thread-terminate!` on this platform is
limited to the current thread (and threads which are not in a running
state).

Implementation of a compatibility layer, implies that we need
to keep a list of all the mutexes owned by the thread to be killed.
This can be done but, since it implies a cost, even on OS with
`pthread_cancel` support, this solution has not been choose.