Don't update the git-info.h file if not necessary
Since a new global link of the VM is not necessary if the newly
generated git-config.h is identical to the previous one, the new
version of the generation script will produce a new version in a
temporary file and if a change is detected, the new version replaces
the old one. Hence, git-config.h is now a standard dependence of
misc.c and a new link is done iff there are changes in the information
that should be displayed by the -V option.