Use an env var for strip
When cross-compiling, the strip binary will not work on the stklos
executable (binary format mismatch). One solution is to use a STRIP
variable, and set it to "stklos" if it is either missing or empty.

Tested on Debian GNU/Linux on amd64, and on the OpenWRT build root
(for MIPS).

This preserves the install-base-no-strip target, which could also
be used for those who want to debug STklos.