Add SRFI-46: Basic Syntax-rules Extensions
STklos already has half of this SRFI implemented: mbe.stk
does understand the tail patterns defined in the SRFI text.
The only missing part was the configurable ellipsis symbol,
which this commit brings.

Since STklos does not recognize symbols beginning with ":",
we check for symbols or keywords in front of the SYNTAX-RULES
symbol. This seems to work, and does not break any tests.

This also includes some basic tests.