Changed the way to make srfis that are implemented in C+Scheme
Instead of having two files installed (a compiled Scheme byte-code file
and a shared library), we embed now the byte-code in the shared
library.

The advantage is that file loading will be faster and we will not have
the files spread in different directory.

The implementation files can be stored in a unique directory under
./lib.

This commit implements SRFI-170 with this new "framework".