Be more precise when turning rationals into doubles
The problem was spotted by @jpellegrini in PR #223.

The PR fix uses the function `mpq_get_d` which is in in GMP but not in
mini-gmp (it is in mini-mpq.c, which is not in included in our gmp-lite
library).

Instead of adding all the mpq library, we adapt the `mpq_get_d` found in
mini_mpq for converting precisely rationals made of bignums in doubles.