Consider GC warnings as errors.
When the first GC warning occurs, we signal a STk error to avoid the complete
operation. Most GC warnings are real errors (mainly no more memory); some are
real warnings but they seem to be rare (never seen them, in fact).

This solution avoids testing if the value returned by the GC
allocation function is NULL, and it should be faster. We could use
real function if this is not convenient.