As before, C compiled files contains module information.
This information was no more produced since commit b02e5d7.
The main reason, was that module information was not used.

However,
 - it breaks the usage of old ScmPkg packages.
 - it could be useful for external tools to find the symbols/syntax defined
   in mixed modules
 - the string whch contain the information is generally not too big

Consequently, the production of module information is reintroduced with this
commit.

As a consequence, mixed-code modules have to add a `DEFINE_MODULE_INFO` at the
end of the C file (to avoid a C compilation warning).