Big review of all Makefiles
All the Makefile.am have been reviewed and cleaned. Now, make
uninstall cleans all the files and directories created. The only
directories which are not removed are bin include lib and share (they
could have been created if the prefix is not standard, but removing
them, probably being root, is too dangerous).