Add lexically-scoped, non-hygienic macros
Two mechanisms are added:

* `%define-syntax`, which defines global macros that are
  only visible in the module where they are created;
* `%let-syntax`, which defines local macros.

This commit doe NOT yet allow one to use internal defines as
if they were `%let-syntax`. This would probably require us to
change the definition of `%define-syntax` from a macro (it is
defined with `define-macro`) to a compiler special form. Not
hard, but not yet done.

The `compile` function is wrapped in a call to `%%macroexpand`,
so macros -- local and global -- are expanded before anything
happens.

The changes to the compiler were kept to the minimum required.

There are some tests included.

Below is a more detailed description of the changes.

ENVIRONMENT:

The representation of the environment by the compiler has been
augmented: it used to be a list of lists of symbols, where each
inner list represented one level of lexical scope.
It is now essentially the same, except that instead of a symbol,
we may also store a pair there:
- `x`         (a symbol) means `x` is bound in the lexical level
- `(x . obj)` (a pair) means `x` is bound to a lexically-scoped
  macro, and `obj` is a `<syntax>` structure.

So we make symbol-in-env? return elements, and rename it.
Symbol-in-env? (which returned either #t or #f) now is a
procedure that returns:
- #f
- the symbol searched (if it's not a macro)
- a pair containing a macro name and its expander

A question that could arise is: why not represent all identifiers as pairs?
Answering: that would require several changes to the compiler. A less invasive
approach was used: wherever (symbol-in-env? ...) was used, the
new function works without any further adaptation on the compiler.

GLOBAL MACROS

A global macro needs to be `DEFINE`d first, so it will have been
computed and can be stored in a global variable. So its value will
be a `<syntax>` structure, containing the expander.
When the compiler finds a global variable of this type, it just
expands and compiles the expanded form.

Scheme implementation usually do not allow `SET!`-ing a variable that
is bound to a macro transformer, but it may be interesting to be
able to change the transformer of a macro.

When the compiler finds a global macro being used,
```
(define-syntax f ...)
...
(f ...)   <= f is bound to a <syntax> structure!
```
then it will call `f`'s expander on that form and call itself
recursively.

The new macros are stored *as ordinary variables*, so they belong to
modules, and can be exported as module symbols.

LOCAL MACROS

A local macro needs to have its identifier inserted in the same
stack structure that holds the names of variables, so lexical
scope works as intended (identifiers for variables -- used at execution
time -- and macros -- expanded at compile time -- may shadow each other).
Also, local macros defined in outermost code should be available for use
in macro code in innermost code:
```
(define-syntax h ...)

(let-syntax ((g ...))
  ...
    (let-syntax ((f ... ,(g ...) ...))  <= f may refer to g and h
```
The local macro name is kept on the environment stack by the compiler
as a pair (not a symbol), so that it's clear it's a macro. The pair
is `(name macro-obj)`, where macro-obj is an instance of `<syntax>`

When the compiler finds a reference to an identifier,
- if it is stored as a symbol, compile reference-to-variable
- if it is stored as a pair, run the expander (which is stored
  along with the macro name), then recursively run the compiler.

HOW EXPANDERS ARE COMPILED

A macro expander should use the environment in which it was defined.
So, `compiled-expander` should be compiled as
```
(eval expander module)
```
When `f` makes a reference to a macro not in its local environment,
then the global environment should be used. But since we have modules,
the global environment is not unique, so the expander is compiled in
the module where the macro was defined.

References to other local macros are dealt with when expanding: the
compiler expands the `car` of a form and starts over, with the same environment.

If a global macro has been defined, then it is available as the value
of a global variable, _in a module_. So, since a local macro must
have access to that global macro, we computed the expander calling
`eval` in it, *on that module* (as mentioned before).

Caveat: it is possible to change the content of the global macro
after the local macro has been defined, but before it has been
used.

HYGIENE/RENAMING

We _DO NOT_ yet deal with those issues here. They can be dealt with
separatelty later, maybe with the help of an `ALIAS` compiler macro.

REMAINING ISSUES

This doesn't seem to work for compiled files. If we compile a
file that creates and then uses a macro, STklos fails to expand it.