Optimize do loops
Do not translate do into letrec.

 (do inits test body)

 --- is rewritten to --->

 (let inits
   (%%label L)
   (if test
       result
       (begin body
              (set! var-1 update-1)
              (set! var-2 update-2)
              ...
              (%%goto L))))

This saves a procedure call in each iteration, and generates less
opcodes. The speedup is considerable:

(time
 (let ((sum 0))
   (do ((i 0 (+ i 1)))
       ((= i 100_000_000) sum)
     (set! sum (+ sum 1)))))

From 7221 with LETREC to 5614 with GOTO.
(NOT a real benchmark; just a quick test -- I've ran 10 times that
code with each do-translator, and those numbers are the average.

The number of opcodes also went from 32 to 20.

This commit also adds sanity checks to the do form:

* too many arguments in each binding (only <2 was checked before)

* number of elements in the test-and-result form