Going back on the implementation of flexvector (using struct)
First implementation of @jpellegrini used structures and it was
replaced by an implementation using objects. The main reason was to
permit that flexvectors can be read back using a #, external form.

However, struct creation and field accesses are largely faster than
their object counterpart, and since the way structures can be printed
can be changed with struct-type-change-writer!.