Suppressed alternative gnu-glob.{c,h} implementation
The suppressed code contained a vulnerability, and it is hard to find a more
recent version that will not drain a lot of dependencies.

In fact, glob is present in POSIX for twenty years now.  So, it should not be
a problem in fact. Glob is present on main Unixes, and Android.  Not having an
alternative implementation could be a problem Win32 (not WSL). In this case
the Scheme glob function returns nil.