Changed the order importations are done in an import
This change is INCOMPATIBLE with previous versions of STklos.

Previously, when we did an `(import A B)`, all the symbols of A
are visible even if they were also defined in B.

Now, to be compatible with the R7RS import directive implementations I have
tested (R7RS seems to be silent on import conflicts), the order is reversed.

Documentation has also been updated. By luck, the documentation of partial
importation was buggy and didn't appear in the final manual reference.

Since the syntax for partial imports is not compatible with the define-library
one (and since it is finally not documented) support for the old form can
probably be deleted (this is not the case yet).