Transitional version of module (used to bootsrap)
This version serves just for prepare the bootstrapping of the new import
syntax (the R7RS one, which only works for libraries for now).

Since this is temporary code, this is very sub-optimal. In particular, old
partial import don't work anymore (but not used when bootstrapping),
and Scmpkg libraries are probably broken.