Implement some algebraic optimization
Don't operate on identities in two-args + and *

(+   0 a) => a
(fx+ 0 a) => a
(*   1 a) => a
(fx* 1 a) => a
(/   a 1) => a

It seems that the `-` case is not needed (STklos already seems to
rewrite it nicely).

The n-args case would need to be handled in code generation, though...

Passes all tests.

Also optimize division