Add non-mutable bindings in modules
Adds the following primitives:

module-lock-define!     disallows define in a module
module-define-allowed?  checks if define is allowed

module-lock-binding!    makes one binding non-mutable
module-lock-bindings!   makes all module bindings non-mutable
binding-mutable?        checks if a binding is mutable

The SCHEME module does not allow new bindings do be DEFINEd,
and does not allow its bindings to be SET!, so it now can be
safely used as a reference for the original Scheme bindings.