Use VM instructions for incremetation/decrementation
The instructions IN_INCR and IN_DECR were already present in the VM
but never generated(!!).

Now adding one or subtracting one will uses these instructions.

Note that the gain in time is really marginal (before we generated
IN_SINT_ADD2 which is mostly as efficient as IN_INCR).

Before: (+ a 1) ==> 000:  GLOBAL-REF           0
                    002:  IN-SINT-ADD2         1

Now: (+ a 1) ==>    000:  GLOBAL-REF           0
                    002:  IN-INCR

The advantage is so minimal that the fx+ and fx- have not been
implemented