Differentiate "not a number" and "not a real number"
The error message when we tried to apply abs, floor,
positive? and similar procedures on a complex number was
"bad number", the same message as if it were a string,
symbol or anything not numerical. We can be more precise
in the error message, without performance penalty and
simplifying a tiny bit the file `number.c`.

This is what weget after this patch is applied:

```
stklos> (abs -1+2i)
**** Error:
abs: `-1+2i' is not a real number
	(type ",help" for more information)
stklos> (abs 'a)
**** Error:
abs: `a' is a bad number
	(type ",help" for more information)
stklos> (positive? -1+2i)
**** Error:
positive?: `-1+2i' is not a real number
	(type ",help" for more information)
stklos> (positive? "x")
**** Error:
positive?: `"x"' is a bad number
	(type ",help" for more information)
stklos> (floor "x")
**** Error:
floor: `"x"' is a bad number
	(type ",help" for more information)
stklos> (floor 2-1i)
**** Error:
floor: `2-1i' is not a real number
	(type ",help" for more information)
```