Added module (srklos markdown env)
This module contains the variables accessible from the document
environment. New bindings can be added with defines embedded in the document
or with metadata definitions.

Now If, Include, System are command treated internally, and no more functions
in the document environment. BTW, the logic of If/Else/End must be rewritten.