Use fixed-size integers for opcodes
Not for everything in STklos -- using "long" is good, since
it makes STklos use the ideal integer size for integer arithmetic
depending on the platform. But the VM will now always use exactly
two bytes for opcodes, so we can be sure that the bytecode will
always be portable.

short          -> int16_t
unsigned short -> uint16_t