Temporary re-implementation of define-macro (with defmacro)
defmacro (whih will be renamed define-macro when finished) is simpler
than the old implementation of macros. It permits the definition of
global macros which can be exported or not (old macros are always
exported when interactive and must be exported with the special
export-syntax form). defmacro macros are exported with the normal R7RS
export form.

The injection of macros in .ostk file is not yet here.