Try to optimize procedure calls
listify_top gets the arguments from the stack and creates a
list. This is done by consing each element, and that means
an allocation call for each.

We can allocate the memory for all arguments at once.

Advantages:
- A single call to STk_cons (and libgc) instead of several calls
- The artuments will stay close to each other, good for caching.
  I guess they would already if libgc was called sequentially to
  allocate them, but "what if" we have two threads making procedure
  calls?

Drawback:
- so long as there is a reference to *one* of these objects, the
whole list will stay in memory and won't be garbage collected.

We allcoate an array res, of cons cell, each pointing to the next:

    +------------+-------------+         +------------+
    |  _____ ___ |   _____ ___ |         |  _____ ___ |
    | | car | ----->| car | ------> ...  | | car | -------> nil
    |  ----- --- |   ----- --- |         |  ----- --- |
    +------------+-------------+         +------------+
         res[0]      res[1]         ...     res[n-1]

And return a pointer to the first one ( &res[0] ).