Optimize vector sorting with fx<, fx>
We have fast paths for sorting vectors when the predicate is either
fx< or fx>. When the vector sorting procedures detect that the
predicate is one of those, it uses plain C "<" and ">" instead.

Remark 1: Fast paths for fx<= and fx>= would be possible, but I'm not
sure many people would use those, and they'd clutter the code.

Remark 2: Fast paths for fl< and fl> would certainly be possible, but
those are not in  STklos by default (it's a SRFI), so things would
become more complicated.