Add SRFI 43: Vector Library
This SRFI was being considered to inclusion in R7RS large when
it was found that it was actually incompatible with R7RS small.
Then SRFI 133 was designed, but it did not officially supersede
SRFI 43.

Since there are only 6 icompatible procedures, and it's easy to
rename symbols when importng and exporting R7RS libraries, there
is no harm in having SFI 43 available also.

This implementation:

* Takes as many procedures as possible from the SCHEME module;
* Takes others from SRFI-133;
* And finally, provides code for the 6 incompatible procedures,
  vector-map, vector-map!, vector-for-each, vector-fold,
  vector-fold-right, vector-count.

NOTE: the tests are not included in `lib/tests/srfis/` yet