Add missing bindings in SRFI-133
when an exported symbol is already defined in STklos, we
actually define it inside the module, as in

(define vector-fill! (symbol-value 'vector-fill! scheme))

because otherwise SRFI-133 would require the SCHEME (or STklos)
module in order to work. Also, if someone changed the binding of
vector-fill!, for example, in module STklos, and SRFI-133 counts on
its original binding, then SRFI-133 will not behave as intended.