Fix vector-reverse-copy! in SRFI-133
R7RS specifies that, for vector-copy!,

"The order in which elements are copied is unspecified, except that if
 the source and destination overlap, copying takes place as if the
 source is first copied into a temporary vector and then into the
 destination. This can be achieved without allocating storage by
 making sure to copy in the correct direction in such circumstances.",

and that vector-reverse-copy! is "Like vector-copy!, but the elements
appear in to in reverse order."

So that restriction is also true for vector-reverse-copy!.

However, I know of no way to do that without allocating extra space
for REVERSE copying of overlapping parts of the same vector.

We allocate new space *of the size of the part being copied*, so if
the vecor has a billion elements and only 10 are being copied, then we
allocate 10 cells.

We split this in two cases, overlapping and non-overlapping...

Also:

- two tests, for this situation, were added to test section for this
  SRFI;
- the Scheme wrapper was rejecting the case when the source and
  target overlaps, but this is implicitly allowed by the spec, so this
  check was removed.