Add the 'environment' procedure
Described in section 6.12 of R7RS.

(environment [name] lib1 lib2 ...)

This procedure returns a specifier for the environment that
results by starting with an empty environment and then
importing each lib (list or symbol), considered as an import
set, into it. The bindings of the environment represented by
the specifier, as is the environment itself.

In STklos,
- each |lib| argument can be a list (specifying an R7RS
  library) or a symbol (specifying a module).
- the return environment is an R7RS library (which can be
  passed to |eval|).
- if |name| is a string, it will be set as the new environment
  (library) name.