Add SRFI-160: Homogeneous numeric vector libraries
SUMMARY:
========

Adds SRFI-160, which extends SRFI 4, bringing ideas from
SRFIs 152 and 133.

The SRFI requires the creation of the libraries:

- (srfi 160 base),  with basic procedures
- (srfi 160 s8),
  (srfi 160 u8),
  (srfi 160 s16),
  ...,
  (srfi 160 c128), with more specific procedures

The modules are named srfi/160/base, srfi/160/s8, and so
on.

- A srfi/160 convenience module is also provided.

Since the motivation for someone to use uniform vectors is
likely to get shorter execution times, most of the procedures
are implemented in C in 160/base.c.

The procedures in base.c are selectively exported. The C
functions accept an integer as type tag, and hence work
for all uniform vectors.

CHANGES IN OTHER PARTS OF STKLOS:
=================================

* reader.c: now accepts #c64 and #c128 syntax for uniform
  complex vectors

* stklos.h: UVECT_C64 and UVECT_C128 are defined

* uvector.c:
  - vector_element_size uvector_set uvector_ref makeuvect
    now handle uniform complex vectors. Complexes are stored
    with their real and imaginary parts together
  - UVECT_C64 and UVECT_C128 are used;
  - vector_element_size uvector_set uvector_ref makeuvect
    are not static anymore, since they're used in base.c
  - Cmake_complex was copied from number.c
  - an EXTERN_PRIMITIVE for exact->inexact was added
  - now includes <float.h> in order to use FLT_MAX.

* lib/srfi/Makefile.am, configure.ac were adjusted so
  that the Makefile in lib/srfi/ recursively goes
  into lib/srfi/160, *and* cleans it up when we call
  Make clean.

SRFI installation:
==================

The SRFI defines several libraries (bot not "(srfi 160)"):

(srfi 160 base)
(srfi 160 s8)
(srfi 160 u8)
...
(srfi 160 c128)

So we need a directory structure like

srfi/160
srfi/160/base.c
srfi/160/base.stk
srfi/160/tagvector-template.stk
srfi/160/Makefile.am

That automake will later install to

$PREFIX/lib/base.so
$PREFIX/share/base.ostk
$PREFIX/share/s8.ostk
$PREFIX/share/u8.ostk
$PREFIX/share/s16.ostk
...
$PREFIX/share/c128.ostk

For this to work, we

- created the 160 directory under srfi/ (of course!)
- added SUBDIRS to lib/srfi/Makefile.am
- also added lib/srfi/160/Makefile to AC_CONFIG_FILES
  in configure.ac.

The files are installed in their proper places.

CONVENIENCE SRFI-160 MODULE:
============================

We also add a convenience srfi/160 module, so the user
can "(import (srfi 160))" and have all the SRFI-160
libraries loaded.

Since "srfi/160.stk" was not being used by the rest of
the implementation, this module was implemented there.

Technically, this is interesting:

The original TAG.stk files are generated from the
tagvector-template.stk file.

tagvector-template.stk
  |
   \_ s8.stk
   |_ u8.stk
   ...
   |_ c128.stk

Each one of those exports a large quantity of symbols, so
having them all in exports clauses in a single file would
be somewhat clumsy.

We do this: srfi/160 will

- import srfi/160/base and all srfi/160/TAG modules
- export the same symbols as srfi/160/base
- from the template export list in tagvector-template.stk,
  a new list of symbols is obtained, and we programatically
  go through it, adding it to the module's export list.

TESTS:
======

The tests from the reference implementation were
adapted for STklos. They need the (srfi 160) module,
which is not part of the spec, but was added for
convenience.