Enhance error message in STk_values2vector
For example, suppose if we write, in C, a procedure similar
to map or fold, which iterates through a structure, applying
a procedure KONS (or PROC).
Also suppose PROC takes 2 arguments **and produces 2 values**.
In C, we'd do

STk_values2vector ( STk_C_apply(PROC, 2, a, b), res );

However, if later the user passed a procedure that produces the
wrong quantity of values, then STkvalues2vector (as it is now)
would signal an error with a cryptic message, "bad vector".
(Indeed, it's a vector of the wrong size, which was passed to
hold the values -- but the Scheme programmer may not even know
that this is how things are implemented internally).

Instead, this patch makes STkvalues2vector raise an error with
a more informative message,

"expected ~S values, but ~S were given"

when the vector size is wrong.

If a non-vector is passed, then we still throw the "bad vector"
message, because 1) it is actually correct, and 2) it represents
an error from the C programmer, who would likely know what happened.