autoload: delete wrapper after autoloading
Autoload was keeping a wrapper over the procedures to be
loaded, and this was slowing down the execution in some
cases. This changes the autoload macro to do remember
if the file was really loaded (if the binding changed)
and issue a "(set! name binding)" *outside* the
definition of the wrapper.

It will somewhat clutter the current-module with gensym'd
symbols that are used to remember if the file is loaded,
but then:

- Maybe it would be really better to use the condition
  system to autoload (but it's more complex);
- Or, maybe we could have an internal %unbind prinitive
  to forget bindings of symbols?