Fix exact-integer-sqrt
In the stop condition of the Newton-Raphson method, the
implementation was checking if xn was equal to xn1. It
should be (>= xn1 xn), and xn should be returned. In one
case, the method oscillated between xn > xn1 and xn < xn1,
never stopping (this happens when n+1 is a perfect square).

While we're at it, the initial guess used for xn was 1.
We now use a better estimate,

(expt 2 (+ 1 (quotient (integer-length n) 2)))

which is the least power of two bigger than SQRT(n).

For that we used the new integer-length primitive.