Changed another time the syntax of the front matter
The syntax for metadata was "(keyword: sexpr)+"
Using a sexpr is not convenient since it is not compatible with the
one used generally (often Yaml). It doesn't work for Emacs
fontification, Emacs deft command, the Pelican static site
generator...

Consequently, the characters after the keyword and up to the end of
line are collected in a (trimmed) string.