compfile.stk is no more compiled in a separate file
Now the function compile-file function is part of the core. The
compilation in an external file was a nightmare with `make -j` as seen
in Issue #427.

The reason to have compile-file in a separate file was to minimize the
size of `boot.img`. A long time ago, `boot.img` was small, but this is
no more the case (the inclusion of compile-file in the core makes the
image 1kb larger, on a 220kb basis).