Optimize empty lambdas
We transform ((lambda () ...)) into (begin ...).

If we're applying a lambda without arguments, ((lambda () ...)), we
may transform it into (begin ...), which is much more efficient. Given
how common it is to pass chunks to procedures, this can speed up
execution considerably:

```
(disassemble-expr '((lambda () a b)))

000:  PREPARE-CALL             <==
001:  ENTER-LET            0   <==
003:  GLOBAL-REF           0
005:  GLOBAL-REF           1
007:  LEAVE-LET                <==
```

The PREPARE-CALL, ENTER-LET and LEAVE-LET instructions are not
necessary, and will not be generated if we actually transform the
lambda into a begin.

We do not transform if the CDDR to be null, since (lambda ()) is not
allowed.

However, this fails:

```
(require "full-syntax")
(let-syntax ((foo (syntax-rules ()
                    ((_ expr) (+ expr 1)))))
  (let ((+ *))
    (foo 3)))
```

expected 4 but got 3

Likely because "full-syntax" was not provided.

Also two other failure in the test suite, but I can't reproduce them:

test Test #3.1 on (let-syntax ((foo (syntax-rules () ((_ expr) (+ expr 1))))) (let ((+ *)) (foo 3))) expected 4 but got 3
test Test #3.2 on (let-syntax ((foo (syntax-rules () ((_ var) (define var 1))))) (let ((x 2)) (begin (define foo +)) (cond (else (foo x))) x)) expected 2 but got 1
test Test #8.1 on (let - ((n (- 1))) n) expected -1 but got 1