Use the procedure plist for keeping its documentation
As for source and formals, it permit to procedures to use less space.
Finally we have saved 3 pointers for procedures (and most of the times
those pointers would be probably STk_nil).

Of course, this is less efficient at runtime, but it will be probably not
frequently used and always interactively.