Use GMP for bignum GCD
If at least one of the arguments is a bignum, call the GMP.

This brings a really nice speedup to gcd *and* to lcm (which
uses gcd internally):

1. GCD bignum x bignum   4.960 -> 0.175
2. LCM bignum x bignum   4.771 -> 0.330
3. GCD bignum x fixnum   7.490 -> 2.106
4. LCM bignum x fixnum  11.648 -> 6.681

1: calculate GCD of two bignums 100_000 times
2: calculate LCM of the same two bignums from (1), 100_000 times
3: calculate GCD of a bignum and a fixnum 10_000_000 times
4: calculate LCM of the same two numbers from (3), 10_000_000 times

Each benchmark was repeated 20 times.

CAVEAT I
========
I did not test with different numbers, and this *may*
make a difference (the GCD can be calculated quicker, depending
on its arguments).

We also deal with some specific cases:

- If an inexact argument is given which does not fit a fixnum,
  we turn it into a bignum and later try to conver the result to
  inexact
- If an inexact argument is given, we give an inexact result.
  BUT, if the result of gcd does not fir a double, we revert to
  an exact bignum.

CAVEAT II
=========

This:

(gcd  73786976294838206473   4)  =>  1
(gcd  73786976294838206473.0 4)  =>  4.0

happens with almost all Schemes I tested (Gauche, Chez, Kawa,
Sagittarius, Gambit, Chibi).

It is because in all of them (and in STklos),

(exact 73786976294838206473.0)
is NOT 73786976294838206473 -- due to imprecision in the
conversion from double into integer.