Keep track of allocated bytes
This patch:

- Adds two fields to the VM structure, allocations and bytes_allocated.
  These are initialized when the thread starts.

- Changes the standard STk_must_malloc and STk_must_malloc_atomic to
  increment those two counters for the current thread

- Adds a %thread-allocations primitive that will return three values,
  the number of allocated bytes and the number of allocation calls
  since the thread started

- Adds a %thread-allocations-reset! primitive that zeroes the two
  counters

A helper function thread_inc_allocs(SCM thr, int size) was added to
vm.c

Example:

```
stklos> (%thread-allocations (current-thread))
2324467
290558
87388
```

In this case, 5743784 bytes (290558 SCM objects) were allocated in
219390 since the start of the thread.

Caveat I: the number of allocations is kept in an unsigned long int,
and there is no provision for recovering from overflow (it will
wrap around).

Caveat II: reallocations are not computed, since it would be a bit
more complicated.