Get rid of boot.img in the git
Boot.img and boot.c are necessary to bootstap STklos. However,
boot.img is a huge file with very long lines. Since, it is which is
often rebuild, it is a real problem with git (particularly when
merging branches with important changes). However, since both files
contain the same information, the first compilation of the build of a
new image uses the image embedded in the executable (instead of the
external boot.img).