Don't define _XOPEN_SOURCE on Darwin 10
Once upon a time it was necessary to define _XOPEN_SOURCE to build gc on
Darwin 10 as the MacPorts project did back when the gc version was 7.1
in macports/macports-ports@bfe12f56 and on Darwin 11 in
macports/macports-ports@78d0aec5, but MacPorts was able to stop doing
that once gc was at version 8.0.4 in macports/macports-ports@1451cea0.

The first instance of adding _XOPEN_SOURCE wasn't correct because it was
setting GCINC to "GCINC -D_XOPEN_SOURCE" rather than (as was probably
intended, and as the code used to do before being refactored in
67ea10f7) appending "-D_XOPEN_SOURCE" to the existing value of GCINC.
But GCINC isn't set at this point in the script anyway and this wrong
value is overwritten later. And this entire section of code is never
reached because OS_NAME_VERSION is not set at this point in the script.