Don't (attempt to) add include path on Darwin 10
The default MacPorts include path was added back in f3583a7c "to support
ScmPkg on MacOs MacPorts" but it would not have been correct if the user
had installed MacPorts in a different prefix or if the user were using
Homebrew instead of MacPorts or if the user had installed the headers
manually.

The configure script uses pkg-config to find most dependencies (except
gmp) wherever they might be installed so adding include paths manually
should not be necessary. If adding include paths were necessary, it
would have been necessary on all Darwin versions, not just 10, and
adding library paths would also have been necessary, which the script
did not do.

This code wasn't actually taking effect anymore because OS_NAME_VERSION
was not set at this point in the script.