Recognize unknown OSes; add special Solaris flags
This reverts part of 67ea10f7 so that the case statement that adds
special flags for Solaris and exits and prints a message on unknown
operating systems once again takes effect. The OS variable is defined by
the shlib-options script run before the case statement whereas
OS_NAME_VERSION is not defined until later in the script.