Remove output file if compilation fails
If anything goes wrong during compilation, print the error or condition
message, remove the output file and exit with a -1 status.

Test: create a file `wrong.stk`,

```
(import nothing)
(define x really-nothing)
```

And compile:

```
stklos-compile -o wrong wrong.stk
```

The error message and the backtrace will be shown, *and* the output
file `wrong` will be deleted.