Optimize dotimes and repeat, and do loop unrolling
1. DOTIMES

dotimes creates a LET even when not necessary:

(macro-expand '(dotimes (i 10) 'work))
  => (let ((G1 10))
       (do ((i 0 (+ i 1)))
           ((>= i G1)
            (void))
         'work))

The new variable G1 is not necessary -- since 10 is constant,
it could have been used in the body instead of G1, and the LET
is superfluous. For nested dotimes, this can make a difference.

With this patch we check if `count` is a number, and when it
is, we generate a begin instead of a let.

For example:
(macro-expand '(dotimes (i MAX) 'work))
    => (let ((G77 MAX)) (do ((i 0 (+ i 1))) ((>= i G77) (void)) 'work))

(macro-expand '(dotimes (i 10) 'work))
    => (begin (do ((i 0 (fx+ i 1))) ((fx>= i 10) (void)) 'work))

The second case will be faster for nested DOTIMEs, because it
avoids access to a local variable *and* entering-and-leaving
a LET.

2. REPEAT

We do loop unrolling! For `repeat` this is very easy, since it does
not use variables that can be changed by the user.

* The parameter `compiler:unroll-iterations` controls the number
  of iterations to unroll
* We have an internal function `%multiply-list` that is used to make
  copies of the body:
  `(%multiply-list '(x y) 3) => (x y x y x y)`
* `repeat` will divide the repetition count by compiler:unroll-iterations,
  and produce `(%repeat ,quot ,(multiply-list body))` followed by
  the remainder times the body:

(macro-expand '(repeat 5 a b c))
=> (begin
     (%repeat 1
       a b c
       a b c
       a b c) ;; 5 / 3 is one, so one iteration of three copies of body
     a b c
     a b c) ;; The remainder of 5 / 3 is two: these are those two copies.

Possible issue: we use list-copy to make the copies of the body. This
*could* maybe have an undesirable effect? Maybe one would expcet
`(eq? x '(...))` to be true, and it won't anymore? Should we not use
list-copy?

3. Both REPEAT *and* DOTIMES

If count (or expr) is constant *and* a fixnum, we can use
fx- and fx>, even though the speed difference is not
that great.