Portable, simpler and faster decode-float + encode-float
1. DECODE-FLOAT

`decode-float`, as implemented currently, will not work on at least
one machine on which I have tried (this could be due to endianness
or to the other machine not implementing IEEE doubles as expected).

This patch implements a new version of this procedure, using
something similar to what ECL does for the Common Lisp function
`integer-decode-float`.

It seems also that this version is faster than the previous one:

```
(let ((a 0))
  (time
    (repeat 10_000_000
      (set! a (decode-float 1.0001))))
  a)
```
With the current version in STklos, the average time to run this
on my machine is 1151.116 ms.
With the new version, it is 882.717 ms.

Also, the code does deal correctly with subnormal numbers (which
ECL does not -- I have changed it a little bit).

It also returns `#t` and `#f` in the significand for NaNs and
infinites, as the previous version did.

As a bonus, the code is now simpler.

2. ENCODE-FLOAT

Inspired bu Gauche, I also implemented `encode-float`, and three
auxiliary procedures:

- `float-max-significand`
- `float-min-exponent`
- `float-max-exponent`

These will give the range of possible values to be used when
calling `encode-float`.

The `encode-float` function also interprets  `#t` and `#f` in
the significand to produce NaNs and infinities.

3. ENCODE-FLOAT AND DECODE-FLOAT TESTS

I have included a number of tests...